landing page refactor
