+++ src/assets/images/index.js
@@ -1,4 +1,4 @@
 export * from './logo'
-export * from './pictures'
+export * from './landingPictures'
 export { default as gooseLogin } from './Goose_login.png'
 export { default as gooseRegister } from './Goose_register.png'
