+++ src/assets/images/landingPictures/index.js
