+++ src/localization/strings/en.js
@@ -1,14 +1,7 @@
 export const en = {
-  "Sign Up": "Sign Up",
-  "Cancel": "Cancel",
-  "Calendar": "Calendar",
-  "Sidebar": "Sidebar",
-  "View": "View",
   "All in": "All in",
-  "One": "One",
-  "GooseTrack's Calendar view provides a comprehensive overview of your schedule, displaying all your tasks, events, and appointments in a visually appealing and intuitive layout.": "GooseTrack's Calendar view provides a comprehensive overview of your schedule, displaying all your tasks, events, and appointments in a visually appealing and intuitive layout.",
-  "GooseTrack offers easy access to your account settings, calendar, and filters. The My Account section allows you to manage your profile information and preferences, while the calendar provides a quick and convenient way to view your upcoming events and tasks.": "GooseTrack offers easy access to your account settings, calendar, and filters. The My Account section allows you to manage your profile information and preferences, while the calendar provides a quick and convenient way to view your upcoming events and tasks.",
-  "GooseTrack is an all-in-one productivity tool that helps you stay on top of your tasks, events, and deadlines. Say goodbye to scattered to-do lists and hello to streamlined productivity with GooseTrack.": "GooseTrack is an all-in-one productivity tool that helps you stay on top of your tasks, events, and deadlines. Say goodbye to scattered to-do lists and hello to streamlined productivity with GooseTrack.",
+  "Calendar": "Calendar",
+  "Cancel": "Cancel",
   "Default": "Default",
   "Email": "Email",
   "Enter email": "Enter email",
@@ -16,18 +9,25 @@ export const en = {
   "Enter text": "Enter text",
   "Enter your name": "Enter your name",
   "Feedback": "Feedback",
+  "GooseTrack is an all-in-one productivity tool that helps you stay on top of your tasks, events, and deadlines. Say goodbye to scattered to-do lists and hello to streamlined productivity with GooseTrack.": "GooseTrack is an all-in-one productivity tool that helps you stay on top of your tasks, events, and deadlines. Say goodbye to scattered to-do lists and hello to streamlined productivity with GooseTrack.",
+  "GooseTrack offers easy access to your account settings, calendar, and filters. The My Account section allows you to manage your profile information and preferences, while the calendar provides a quick and convenient way to view your upcoming events and tasks.": "GooseTrack offers easy access to your account settings, calendar, and filters. The My Account section allows you to manage your profile information and preferences, while the calendar provides a quick and convenient way to view your upcoming events and tasks.",
+  "GooseTrack's Calendar view provides a comprehensive overview of your schedule, displaying all your tasks, events, and appointments in a visually appealing and intuitive layout.": "GooseTrack's Calendar view provides a comprehensive overview of your schedule, displaying all your tasks, events, and appointments in a visually appealing and intuitive layout.",
   "Let go of the past and focus on the present!": "Let go of the past and focus on the present!",
   "Log out": "Log out",
   "My Account": "My Account",
   "Name": "Name",
+  "One": "One",
   "Password": "Password",
   "Rating": "Rating",
   "Review": "Review",
   "Save": "Save",
+  "Sidebar": "Sidebar",
+  "Sign Up": "Sign Up",
   "This is a CORRECT email": "This is a CORRECT email",
   "This is a CORRECT password": "This is a CORRECT password",
   "User Pane": "User Pane",
   "User Profile": "User Profile",
+  "View": "View",
   "Well done": "Well done",
   "email is a required field": "email is a required field",
   "email must be a valid email": "email must be a valid email",
