+++ src/navigation/provider.js
@@ -8,20 +8,16 @@ import { BASE_GITHUB_PAGES_URL } from '../utils/constants'
 const LoginPage = lazy(() => import('../pages/auth/LoginPage'))
 const RegisterPage = lazy(() => import('../pages/auth/RegisterPage'))
 const CalendarPage = lazy(() => import('../pages/calendar/CalendarPage'))
-const LandingPage = lazy(() => import('../pages/landing/Landing'))
+const LandingPage = lazy(() => import('../pages/landing/LandingPage'))
 
 export const AppRouterProvider = () => (
   <BrowserRouter basename={BASE_GITHUB_PAGES_URL}>
     <Suspense fallback={<Loader />}>
       <Routes>
-        <Route
-          path={ROUTES.LANDING}
-          element={
-            <Suspense fallback={<Loader />}>
-              <LandingPage />
-            </Suspense>
-          }
-        />
+        <Route path={ROUTES.LOGIN} element={<LoginPage />} />
+        <Route path={ROUTES.REGISTER} element={<RegisterPage />} />
+        <Route path={ROUTES.LANDING} element={<LandingPage />} />
+
         <Route element={<MainLayout />}>
           <Route
             path={ROUTES.PROFILE}
@@ -40,9 +36,6 @@ export const AppRouterProvider = () => (
             }
           />
         </Route>
-
-        <Route path={ROUTES.LOGIN} element={<LoginPage />} />
-        <Route path={ROUTES.REGISTER} element={<RegisterPage />} />
 
         {/* TODO: Add 404 Page */}
         <Route path='*' element={<div>404 Page</div>} />
