+++ src/navigation/routes.js
@@ -1,8 +1,7 @@
 export const ROUTES = {
-  HOME: '/',
   CALENDAR: '/calendar',
   PROFILE: '/profile',
   LOGIN: '/login',
-  LANDING: '/landing',
+  LANDING: '/',
   REGISTER: '/register',
 }
